refactor: 이벤트 리스너 함수 공통화 처리
