+++ src/utils/searchHandler.js
@@ -58,40 +58,3 @@ export const clearSearch = async (state, renderCallback) => {
     console.error("검색 초기화 실패:", error);
   }
 };
-
-// 검색 이벤트 리스너 설정 함수
-export const setupSearchEventListeners = (state, renderCallback) => {
-  const searchInput = document.getElementById("search-input");
-  const searchButton = document.getElementById("search-button");
-  const clearButton = document.getElementById("clear-search-button");
-
-  if (!searchInput) return;
-
-  // 검색 버튼 클릭 이벤트
-  if (searchButton) {
-    searchButton.addEventListener("click", () => {
-      const searchTerm = searchInput.value.trim();
-      if (searchTerm) {
-        performSearch(state, searchTerm, renderCallback);
-      }
-    });
-  }
-
-  // Enter 키 이벤트
-  searchInput.addEventListener("keypress", (e) => {
-    if (e.key === "Enter") {
-      const searchTerm = searchInput.value.trim();
-      if (searchTerm) {
-        performSearch(state, searchTerm, renderCallback);
-      }
-    }
-  });
-
-  // 검색어 초기화 버튼
-  if (clearButton) {
-    clearButton.addEventListener("click", () => {
-      searchInput.value = "";
-      clearSearch(state, renderCallback);
-    });
-  }
-};
